requirements.txt
requests>=2.25.1
pymysql>=1.0.2
pandas>=1.3.0
openpyxl>=3.0.7
schedule>=1.1.0
aiohttp>=3.8.0
keyring>=23.0.0
pycryptodome>=3.15.0